// src/client/matrix/schema/gen_types/KnockedRoom.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface KnockedRoom {
  knock_state?: {
    events?: StrippedStateEvent[];
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface StrippedStateEvent {
  content: {
    [k: string]: unknown;
  };
  sender: string;
  state_key: string;
  type: string;
  [k: string]: unknown;
}