// src/client/matrix/schema/gen_types/InvitedRoom.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface InvitedRoom {
  invite_state?: InviteState;
  [k: string]: unknown;
}
export interface InviteState {
  events?: StrippedStateEvent[];
  [k: string]: unknown;
}
export interface StrippedStateEvent {
  content: {
    [k: string]: unknown;
  };
  sender: string;
  state_key: string;
  type: string;
  [k: string]: unknown;
}