// src/client/matrix/schema/gen_types/ClientEventWithoutRoomID.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ClientEventWithoutRoomID {
  content: {
    [k: string]: unknown;
  };
  event_id: string;
  origin_server_ts: number;
  sender: string;
  state_key?: string;
  type: string;
  unsigned?: {
    age?: number;
    prev_content?: {
      [k: string]: unknown;
    };
    redacted_because?: ClientEventWithoutRoomID1;
    transaction_id?: string;
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface ClientEventWithoutRoomID1 {
  content: {
    [k: string]: unknown;
  };
  event_id: string;
  origin_server_ts: number;
  sender: string;
  state_key?: string;
  type: string;
  unsigned?: {
    age?: number;
    prev_content?: {
      [k: string]: unknown;
    };
    redacted_because?: ClientEventWithoutRoomID1;
    transaction_id?: string;
    [k: string]: unknown;
  };
  [k: string]: unknown;
}