// src/client/matrix/schema/gen_types/m_room_create.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface MRoomCreate {
  creator: string;
  "m.federate"?: boolean;
  predecessor?: {
    event_id: string;
    room_id: string;
    [k: string]: unknown;
  };
  room_version?: string;
  type?: string;
  [k: string]: unknown;
}